generation/carla-simulation/src/client/pyproject.toml
[project]
name = "client"
version = "0.1.0"
description = "CARLA simulator client"
requires-python = ">=3.10"
dependencies = [
    "carla>=0.9.15",
    "joblib>=1.5.2",
    "matplotlib>=3.10.7",
    "notebook>=7.5.0",
    "numpy>=2.2.6",
    "open3d>=0.19.0",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "ply>=3.11",
    "psutil>=7.1.3",
    "pygame>=2.6.1",
    "pyqt5>=5.15.11",
    "pyyaml>=6.0.3",
    "tqdm>=4.67.1",
]
